+++ llama-cpp-bindings/src/model/params.rs
@@ -307,7 +307,9 @@ fn fit_params_status_to_result(
             let message = unsafe { crate::ffi_error_reader::read_and_free_cpp_error(out_error) };
             Err(FitError::Reported { message })
         }
-        other => Err(FitError::UnrecognizedStatusCode { code: other }),
+        other => Err(FitError::UnrecognizedStatusCode {
+            code: i64::from(other),
+        }),
     }
 }
 
@@ -847,7 +849,7 @@ mod tests {
                 std::ptr::null_mut(),
             ),
             Err(crate::error::fit_error::FitError::UnrecognizedStatusCode {
-                code: llama_cpp_bindings_sys::llama_rs_fit_params_status::MAX
+                code: i64::from(llama_cpp_bindings_sys::llama_rs_fit_params_status::MAX)
             }),
         );
     }
